+++ src/datasources/test-plans/components/TestPlansQueryEditor.tsx
@@ -106,6 +106,13 @@ export function TestPlansQueryEditor({ query, onChange, onRunQuery, datasource }
       handleQueryChange({ ...query, queryBy });
     }
   };
+  
+  useEffect(() => {
+    if (!query.outputType) {
+      handleQueryChange({...query, outputType: OutputType.Properties});
+    }
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
     <>