+++ src/datasources/work-orders/components/WorkOrdersQueryEditor.test.tsx
@@ -85,6 +85,31 @@ describe('WorkOrdersQueryEditor', () => {
     expect(take).toHaveDisplayValue('');
   });
 
+  
+  it('should call onRunQuery on init', async() => {
+    const query = {
+      refId: 'A',
+    }
+
+    await renderElement(query);
+
+    expect(mockOnChange).toHaveBeenCalledWith(expect.objectContaining({ outputType: OutputType.Properties, refId: 'A' }));
+    expect(mockOnRunQuery).toHaveBeenCalled();
+  });
+  
+  it('should not call onRunQuery after init', async() => {
+    const query = {
+      refId: 'A',
+      outputType: OutputType.Properties,
+    }
+    jest.clearAllMocks();
+
+    await renderElement(query);
+
+    expect(mockOnChange).not.toHaveBeenCalled();
+    expect(mockOnRunQuery).not.toHaveBeenCalled();
+  });
+
   describe('output type is total count', () => {
     let container: RenderResult;
     beforeEach(async() => {